Revert "dax: fix race between simultaneous faults"
This reverts commit 843172978bb929973 ("dax: fix race between simultaneous faults").

This commit and 46c043ede471 ("mm: take i_mmap_lock in
unmap_mapping_range() for DAX") introduced a number of deadlocks and other
issues, and need to be reverted for the v4.3 kernel.  The list of issues
in DAX after these commits (some newly introduced by the commits, some
preexisting) can be found here:

https://lkml.org/lkml/2015/9/25/602 (Subject: "Re: [PATCH] dax: fix
deadlock in __dax_fault").

Signed-off-by: Ross Zwisler <ross.zwisler@linux.intel.com>
Cc: Alexander Viro <viro@zeniv.linux.org.uk>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Dave Chinner <david@fromorbit.com>
Cc: Jan Kara <jack@suse.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Matthew Wilcox <matthew.r.wilcox@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>