powerpc/mm: add page soft dirty tracking
User space checkpoint and restart tool (CRIU) needs the page's change to
be soft tracked.  This allows to do a pre checkpoint and then dump only
touched pages.

This is done by using a newly assigned PTE bit (_PAGE_SOFT_DIRTY) when the
page is backed in memory, and a new _PAGE_SWP_SOFT_DIRTY bit when the page
is swapped out.

The _PAGE_SWP_SOFT_DIRTY bit is dynamically put after the swap type in the
swap pte.  A check is added to ensure that the bit is not overwritten by
_PAGE_HPTEFLAGS.

Signed-off-by: Laurent Dufour <ldufour@linux.vnet.ibm.com>
Cc: Aneesh Kumar K.V <aneesh.kumar@linux.vnet.ibm.com>
Cc: Pavel Emelyanov <xemul@parallels.com>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>